Refactor: Redis, nginx, Zookeeper, and Kafka plugins to load configurations from shell scripts.
